feat(triton): support fused_slice_v2 with variable-length slices via namedtuple     * Added `fused_slice_v2` Triton op to support variable-length slicing. This extends the original `fused_slice`, which only handled fixed-length slices (even when multiple inputs were supported).
feat(structure): update multiple_slice_cat
    * Added `fused_slice_cat_v2` to replace `fused_slice_cat`. Now, the output of `fused_slice_cat` directly feeds into `fused_slice_v2` instead of multiple `aten.slice` ops.This transforms the original pattern: `fused_slice_cat` + multiple `fused_slice` + multiple `aten.slice` to:  1 * `fused_slice_cat` + 1 * `fused_slice_v2`, reducing ~50 kernels.
    * Added pattern `fuse_multiple_cat2` to merge multiple `fused_slice_cat_v2` ops with the same input into a single call, reducing another 11 kernels.
    * For modelA:  `aten.slice` reduced from 7119 → 1602 `aten.cat` reduced from 209 → 72 . Entire fused pattern (`fused_slice_cat` + `fused_slice_v2`) now takes <4% of runtime.
    * Add embcat in fused_slice_cat_v2.

feat(combosum): add combo sum after fused_slice_cat

+++ xpu_graph/passes/patterns/targets/mlu/triton_kernel/fused_slice_cat.py
@@ -71,4 +71,4 @@ def fused_slice_cat(
 
 @fused_slice_cat.register_fake
 def fused_slice_cat_fake(input_tensor, indices_tensor, n_rows, elements, input_stride):
-    return torch.empty(n_rows, elements, device=input_tensor.device)
+    return torch.empty(n_rows, elements, device=input_tensor.device, dtype=input_tensor.dtype)
